requirements.txt
python-dotenv>=1.0.0
requests==2.32.5
aiohttp>=3.8.0
websockets>=12.0
tenacity>=9.1.2
pydantic>=1.8.0
paradex-py          # 必須手動加入，因為你的代碼依賴它
starknet-py         # Paradex 依賴
